Added a search engine
Created a new form. Then made a new function called search, taking care of searching for the city, updating the UI, making the API call.

That same function is also going to be called by the form whenever it's submitted, and the search is going to be called based on what the user typed in the element.